Finalize box_management legacy utils extraction
