+++ box_management/services/articles/fetch_article_content.py
@@ -1,3 +1,3 @@
-from box_management.domain.legacy_utils import _extract_import_preview_from_url
+from box_management.integrations.article_scraper import _extract_import_preview_from_url
 
 __all__ = ["_extract_import_preview_from_url"]
