--- box_management/views.py
@@ -1,94 +1,0 @@
-"""Compatibility re-export module for box management API views.
-
-Business/API view implementations live under ``box_management.api.views``.
-"""
-
-from .api.views.comments import (
-    ClientAdminCommentListView,
-    ClientAdminCommentModerateView,
-    ClientAdminCommentRestrictionListCreateView,
-    CommentCreateView,
-    CommentDetailView,
-    CommentReportView,
-    DepositRepliesView,
-    ReactionView,
-)
-from .api.views.core import (
-    ActiveBoxSessionsView,
-    BoxSessionView,
-    ClientAdminArticleDetailView,
-    ClientAdminArticleImportPageView,
-    ClientAdminArticleListCreateView,
-    ClientAdminIncitationDetailView,
-    ClientAdminIncitationListCreateView,
-    EconomyConfigView,
-    EmojiCatalogView,
-    GetBox,
-    GetMain,
-    Location,
-    ManageDiscoveredSongs,
-    PinnedSongView,
-    PublicVisibleArticleDetailView,
-    PublicVisibleArticlesView,
-    PurchaseEmojiView,
-    ResolveProviderLinkView,
-    RevealSong,
-    ShareLinkCreateView,
-    ShareLinkPublicDetailView,
-    UserDepositsView,
-    sticker_redirect_view,
-    sticker_root_not_found_view,
-)
-from .api.views.stickers import (
-    ClientAdminStickerAssignView,
-    ClientAdminStickerConfirmDownloadView,
-    ClientAdminStickerDownloadView,
-    ClientAdminStickerGenerateView,
-    ClientAdminStickerInstallView,
-    ClientAdminStickerListView,
-    ClientAdminStickerUnassignView,
-)
-from .services.articles.fetch_article_content import _extract_import_preview_from_url  # noqa: F401
-from .services.deposits.song_creation import create_song_deposit  # noqa: F401
-
-__all__ = [
-    "ActiveBoxSessionsView",
-    "BoxSessionView",
-    "ClientAdminArticleDetailView",
-    "ClientAdminArticleImportPageView",
-    "ClientAdminArticleListCreateView",
-    "ClientAdminCommentListView",
-    "ClientAdminCommentModerateView",
-    "ClientAdminCommentRestrictionListCreateView",
-    "ClientAdminIncitationDetailView",
-    "ClientAdminIncitationListCreateView",
-    "ClientAdminStickerAssignView",
-    "ClientAdminStickerConfirmDownloadView",
-    "ClientAdminStickerDownloadView",
-    "ClientAdminStickerGenerateView",
-    "ClientAdminStickerInstallView",
-    "ClientAdminStickerListView",
-    "ClientAdminStickerUnassignView",
-    "CommentCreateView",
-    "CommentDetailView",
-    "CommentReportView",
-    "DepositRepliesView",
-    "EconomyConfigView",
-    "EmojiCatalogView",
-    "GetBox",
-    "GetMain",
-    "Location",
-    "ManageDiscoveredSongs",
-    "PinnedSongView",
-    "PublicVisibleArticleDetailView",
-    "PublicVisibleArticlesView",
-    "PurchaseEmojiView",
-    "ReactionView",
-    "ResolveProviderLinkView",
-    "RevealSong",
-    "ShareLinkCreateView",
-    "ShareLinkPublicDetailView",
-    "UserDepositsView",
-    "sticker_redirect_view",
-    "sticker_root_not_found_view",
-]
